add encryption tool
