+++ dygraph/deploy/cpp/encryption/include/model_code.h
@@ -1,5 +1,20 @@
-#ifndef PADDLE_MODEL_PROTECT_MODEL_CODE_H
-#define PADDLE_MODEL_PROTECT_MODEL_CODE_H
+//   Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#pragma once
+
+#ifndef DYGRAPH_DEPLOY_CPP_ENCRYPTION_INCLUDE_MODEL_CODE_H_
+#define DYGRAPH_DEPLOY_CPP_ENCRYPTION_INCLUDE_MODEL_CODE_H_
 
 #ifdef __cplusplus
 extern  "C" {
@@ -26,4 +41,4 @@ extern  "C" {
 }
 #endif
 
-#endif //PADDLE_MODEL_PROTECT_MODEL_CODE_H
+#endif  // DYGRAPH_DEPLOY_CPP_ENCRYPTION_INCLUDE_MODEL_CODE_H_
