+++ dygraph/deploy/cpp/encryption/include/paddle_model_decrypt.h
@@ -1,7 +1,22 @@
+//   Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 #pragma once
 
 #include <stdio.h>
-#include "paddle_inference_api.h"
+#include <string>
+
+#include "encryption/include/paddle_inference_api.h"
 
 #ifndef PADDLE_MODEL_PROTECT_API_PADDLE_MODEL_DECRYPT_H
 #define PADDLE_MODEL_PROTECT_API_PADDLE_MODEL_DECRYPT_H
@@ -47,4 +62,4 @@ PDD_MODEL_API std::string decrypt_file(const char* file_path, const char* key);
 }
 #endif
 
-#endif //PADDLE_MODEL_PROTECT_API_PADDLE_MODEL_DECRYPT_H
+#endif  // PADDLE_MODEL_PROTECT_API_PADDLE_MODEL_DECRYPT_H
