+++ dygraph/deploy/cpp/encryption/include/paddle_model_encrypt.h
@@ -1,6 +1,20 @@
+// Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 #pragma once
 
 #include <iostream>
+#include <string>
 
 #ifndef PADDLE_MODEL_PROTECT_API_PADDLE_MODEL_ENCRYPT_H
 #define PADDLE_MODEL_PROTECT_API_PADDLE_MODEL_ENCRYPT_H
@@ -36,7 +50,9 @@ PDE_MODEL_API std::string paddle_generate_random_key();
  * @param dst_dir
  * @return
  */
-PDE_MODEL_API int paddle_encrypt_dir(const char* keydata, const char* src_dir, const char* dst_dir);
+PDE_MODEL_API int paddle_encrypt_dir(const char* keydata,
+                                     const char* src_dir,
+                                     const char* dst_dir);
 
 /**
  * encrypt file
@@ -45,12 +61,16 @@ PDE_MODEL_API int paddle_encrypt_dir(const char* keydata, const char* src_dir, c
  * @param outfile
  * @return          error_code
  */
-PDE_MODEL_API int paddle_encrypt_model(const char* keydata, const char* infile, const char* outfile);
+PDE_MODEL_API int paddle_encrypt_model(const char* keydata,
+                                       const char* infile,
+                                       const char* outfile);
 
-PDE_MODEL_API int encrypt_stream(const std::string &keydata, std::istream &in_stream, std::ostream &out_stream);
+PDE_MODEL_API int encrypt_stream(const std::string &keydata,
+                                 std::istream &in_stream,
+                                 std::ostream &out_stream);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif //PADDLE_MODEL_PROTECT_API_PADDLE_MODEL_ENCRYPT_H
+#endif  // PADDLE_MODEL_PROTECT_API_PADDLE_MODEL_ENCRYPT_H
