--- dygraph/deploy/cpp/encryption/src/util/log.h
@@ -1,9 +1,0 @@
-#ifndef PADDLE_MODEL_PROTECT_UTIL_LOG_H
-#define PADDLE_MODEL_PROTECT_UTIL_LOG_H
-
-#include <stdio.h>
-
-#define LOGD(fmt,...)\
-    printf("{%s:%u}:" fmt "\n", __FUNCTION__, __LINE__,  ##__VA_ARGS__)
-
-#endif //PADDLE_MODEL_PROTECT_UTIL_LOG_H
