+++ dygraph/deploy/cpp/model_deploy/common/include/base_model.h
@@ -24,6 +24,7 @@
 #include "model_deploy/common/include/base_preprocess.h"
 #include "model_deploy/common/include/output_struct.h"
 #include "model_deploy/engine/include/engine.h"
+#include "encryption/include/paddle_model_decrypt.h"
 
 namespace PaddleDeploy {
 
@@ -46,15 +47,15 @@ class Model {
   // Init model_type.
   explicit Model(const std::string model_type) : model_type_(model_type) {}
 
-  virtual bool Init(const std::string& cfg_file) {
-    if (!YamlConfigInit(cfg_file)) return false;
+  virtual bool Init(const std::string& cfg_file, const std::string key = "") {
+    if (!YamlConfigInit(cfg_file, key)) return false;
     if (!PreprocessInit()) return false;
     if (!PostprocessInit()) return false;
     return true;
   }
 
-  virtual bool YamlConfigInit(const std::string& cfg_file) {
-    // YAML::Node yaml_config_ = YAML::LoadFile(cfg_file);
+  virtual bool YamlConfigInit(const std::string& cfg_file,
+                              const std::string key) {
     std::cerr << "Error! The Base Model was incorrectly entered" << std::endl;
     return false;
   }
