+++ dygraph/deploy/cpp/model_deploy/paddlex/include/x_model.h
@@ -35,7 +35,7 @@ class PaddleXModel : public Model {
 
   bool GenerateTransformsConfig(const YAML::Node &preprocess_op);
 
-  bool YamlConfigInit(const std::string &cfg_file);
+  bool YamlConfigInit(const std::string &cfg_file, const std::string key);
 
   bool PreprocessInit();
